Remove python-stats usage
Steal a pearson correlation function implementation from
stackoverflow [1] which itself is taken from Programming
Collective Intelligence which the rest of this code is
based on.

The removal of python-stats is done first because having
a private copy of a library is a bad idea then because
it is GPLv2 which restricts (sigh) its usage with
proprietary applications.

[1] http://stackoverflow.com/questions/3949226/calculating-pearson-correlation-and-significance-in-python